Enable fine-granular filter queries for virtual machines
Instead of providing resource groups in the configuration file, the user
is now able to provide filter queries. The filter queries are passed
locally to the action. With this solution the user is able to use a
fine-granular filter.

Requirement: The resource-graph extension is a pre-requirement. The
application will install the resource-graph extension if is not
installed.

Resolves: #21
Signed-off-by: Auli Rahnasto <auli.rahnasto@gmx.de>